modify root view and added working links to navigate

+++ app/helpers/users_helper.rb
@@ -1,18 +1,12 @@
 module UsersHelper
   def show_edit_user(user)
+    return unless current_user.id == user.id
     link_to 'Edit', edit_user_path(user)
   end
 
   def show_destroy_user(user)
+    return unless current_user.id == user.id
     link_to 'Destroy', user, method: :delete, data: { confirm: 'Are you sure?' }
-  end
-
-  def show_edit_post(post)
-    link_to 'Edit', edit_post_path(post)
-  end
-
-  def show_destroy_post(post)
-    link_to 'Destroy', post, method: :delete, data: { confirm: 'Are you sure?' }
   end
 
   def show_user_link(user)
